Reorganize structure of containers and routing
