+++ app/controllers/all_stocks_controller.rb
@@ -1,0 +1,4 @@
+class AllStocksController < ApplicationController
+  def index
+  end
+end
