+++ app/controllers/api/v1/stocks_controller.rb
@@ -1,5 +1,6 @@
 class Api::V1::StocksController < ApiController
   def index
-    render json: Stock.all
+  end
+  def show
   end
 end
