--- app/controllers/portfolios_controller.rb
@@ -1,4 +1,0 @@
-class PortfoliosController < ApplicationController
-  def index
-  end
-end
