+++ app/javascript/react/containers/stocks/StocksShowContainer.js
@@ -1,0 +1,16 @@
+import React from 'react'
+
+class StocksShowContainer extends React.Component {
+  constructor(props){
+    super(props)
+    this.state = {}
+  }
+
+  render(){
+    return(
+      <div>This is the stocks show page</div>
+    )
+  }
+}
+
+export default StocksShowContainer
