+++ app/views/all_stocks/index.html.erb
