+++ app/views/stocks/show.html.erb
@@ -1,0 +1,1 @@
+<div id="app" />
